requirements.txt
alabaster==0.7.10
Babel==2.4.0
certifi==2017.4.17
chardet==3.0.4
click==6.7
colorama==0.3.9
defusedxml==0.5.0
Django==1.11.2
djangorestframework==3.6.3
docopt==0.6.2
docutils==0.13.1
feedparser==5.2.1
gitsome==0.7.2
gunicorn==19.7.1
idna==2.5
imagesize==0.7.1
Jinja2==2.9.6
jira==1.0.10
MarkupSafe==1.0
numpydoc==0.7.0
oauthlib==2.0.2
pbr==3.1.1
ply==3.10
prompt-toolkit==1.0.14
psycopg2==2.7.1
Pygments==2.2.0
pytz==2016.10
requests==2.18.1
requests-oauthlib==0.8.0
requests-toolbelt==0.8.0
six==1.10.0
snowballstemmer==1.2.1
Sphinx==1.6.2
sphinxcontrib-websupport==1.0.1
uritemplate==3.0.0
uritemplate.py==3.0.2
urllib3==1.21.1
wcwidth==0.1.7
